GPS osaa näyttää kartalta missä pelaaja on. Singleton objekti joka palaa takaisin peleissa ja poistuu pelistä on nyt pelissä toiminnassa.

+++ FlagWorldGame/Assets/Games/GPSTest/Scripts/Location.cs
@@ -7,9 +7,9 @@ using UnityEngine.UI;
 public class Location
 {
     public string name;
-    public float longitude;
-    public float latitude;
-    public float rangeDistance;
+    public double longitude;
+    public double latitude;
+    public double rangeDistance;
     public Image image;
     public string title;
     [TextArea]
